Adds event base mock
